refactor(daemon): group browser bridges into browser/
