+++ apps/daemon/src/browser/browser-open.ts
