+++ apps/daemon/src/diagnostics-export.ts
@@ -27,7 +27,7 @@ import {
 import { readCurrentAppVersionInfo } from './app-version.js';
 import { agentCliEnvForAgent, readAppConfig } from './app-config.js';
 import { spawnEnvForAgent } from './agents.js';
-import { collectBrowserUseDiscoveryFacts } from './browser-use-diagnostics.js';
+import { collectBrowserUseDiscoveryFacts } from './browser/index.js';
 
 interface ResolvedAgentHomes {
   amrOpenCodeHome: string | null;
